Styles System => Stable UWP and Android A UWP project compatible with Shell was added
Remarks:
To display the Custom Icon Font in the "Flyout sector" in UWP you must add the Size and Color properties in Icons.xaml, unfortunately in the "Tabs sector" that logic does not work, that problem must be solved.
In UWP the FontIcons and fonts were added in the "Assets / Fonts" folder, as for the images a good practice is to add them in the root of the UWP project. The FlyoutBackgroundImage property does not work in UWP so I added the image in the FlyoutHeaderTemplate but it is only visible to UWP.